smip/backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.2
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blinker==1.9.0
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.8
comm==0.2.2
comtypes==1.4.1
contexttimer==0.3.3
cryptography==44.0.0
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docker==7.1.0
dxcam==0.0.5
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.4.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
instaloader==4.13.2
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.2.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kinde-python-sdk==1.2.6
kiwisolver==1.4.4
langchain==0.0.279
langsmith==0.0.33
MarkupSafe==3.0.2
marshmallow==3.20.1
matplotlib==3.5.3
matplotlib-inline==0.1.7
missingno==0.5.1
mistune==3.0.2
MouseInfo==0.1.3
multidict==6.0.4
multimethod==1.9
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
notebook==7.1.3
notebook_shim==0.2.4
numexpr==2.8.5
numpy==1.23.4
openai==0.28.0
opencv-python==4.9.0.80
overrides==7.7.0
packaging==24.0
pandas==1.5.1
pandas-profiling==3.4.0
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.3
pexpect==4.9.0
phik==0.12.2
Pillow==9.3.0
platformdirs==4.2.0
playwright==1.49.1
praw==7.8.1
prawcore==2.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==1.10.2
pyee==12.0.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.8.0
PyMsgBox==1.0.9
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytweening==1.2.0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rubicon-objc==0.4.8
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
srsly==2.5.0
stack-data==0.6.3
statsmodels==0.13.5
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.1.0
TikTokApi==7.0.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzlocal==5.2
update-checker==0.18.0
uri-template==1.3.0
urllib3==2.2.3
visions==0.7.5
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
widgetsnbextension==4.0.10
yarl==1.9.2
